split httprout and gateway attched
